+++ cli/docs/changelog.md
@@ -1,16 +1,34 @@
 # TBD
+
+# 0.13.0
+### Features
+* `enclave inspect` now prints port IDs, in sorted order
+
 ### Changes
-* Upgraded `container engine lib` to 0.12.0
-* Upgraded `kurtosis engine lib` to 1.13.1
+* Upgraded `container engine lib` to 0.14.3
+* Upgraded `kurtosis core lib` to 1.42.2
 * Replaced `DockerManager` with `KurtosisBackend` in:
   * `EngineConsumingKurtosisCommand` struct and in `RunFunc`
   * `clean` command
   * `enclave inspect` command
   * `enclave rm` command
-  * `service add` command
   * `service logs` command
-  * `service rm` command
   * `service shell` command
+
+### Fixes
+* Fix a bug where Kurtosis would fail to process containers using the old port spec
+* Fixed a bug where `enclave inspect` crashes if a module is stopped
+
+### Breaking Changes
+* Upgraded `kurtosis engine lib` to 1.14.2
+  * Users should upgrade to `kurtosis-engine-api-lib` 1.14.2
+
+# 0.12.2
+### Fixes
+* Bumping dependencies on:
+    * Engine Server 1.13.2
+    * Container Engine Lib 0.13.0
+    * Core 1.41.1
 
 # 0.12.1
 # Breaking Changes